Refactored task reported
